Extract verifier_diagnostic_flow.rs from turn.rs (parent #680)
Move the Issue #637 / #638 / #654 verifier-diagnostic pass flow out of
`turn.rs` into a focused sibling module so `turn.rs` keeps shrinking
toward responsibility-focused units (parent meta-issue #680).

Three production helpers + three private helpers extracted as free
functions taking `&mut Agent` (matching `actor_loop_flow` /
`reply_retry` / earlier vertical-slice precedent):

- `prepare_verifier_diagnostic_pass` (clears stale state, reserves
  attempt slot, derives `PreparedVerifierDiagnosticPass`, emits
  behavior-contract projection event)
- `request_verifier_diagnostic_reply` (invokes verifier diagnostic
  chat with overridden timeout / max_predict)
- `handle_verifier_diagnostic_failure` (records the failure + decides
  between `RetryPending` and terminal `Unavailable`)
- `record_verifier_diagnostic_failure` (private): store-boundary SSOT
  sanitize + repair-job mutate + bounded snapshot refresh + emit
  `agent.verifier_diagnostic.failed`
- `record_verifier_diagnostic_unavailable` (private): same sanitize /
  snapshot / emit, then forwards to
  `safe_stop_emit::emit_safe_stop_report_for_diagnostic_target_missing`
- `reset_verifier_diagnostic_state_if_needed` (private): clears stale
  assessment after cluster advance / target exhaustion

The `VERIFIER_DIAGNOSTIC_MAX_PREDICT` constant moved with the cluster
(its only caller).

Behavior unchanged: same Issue #637 (CB-002) sanitize pipeline at every
diagnostic_error write boundary, same Issue #638 bounded snapshot
refresh on retry-pending + terminal paths, same Issue #654 cascade
into the safe-stop emit on the terminal path, same
`VERIFIER_DIAGNOSTIC_ATTEMPT_LIMIT` retry budget.

`pub(super)` limited / no facade re-export (DR3-001). Callers updated
in `verifier_orchestration.rs` (4 sites) to the free-fn form.

LOC delta:
- turn.rs: 2,397 → 2,200 (-197)
- new module: +233

3,114 lib tests pass; `cargo fmt` / `cargo clippy --lib --all-targets
-- -D warnings` clean.
